The Just Prince
This is a book about art of war, culture , philosophy and political science which tells how to manage conflicts between states and people.

I have been choosen this name not because I am a dictator or boss but I want to carry the responsibility of The Free Software Movement in another words leading to The Free Software Movement.

I / We design a network and group of people who is responsible for supporting and developing Free Software and I assume that there will be conflicts and this "The Just Prince" title is a sign to show the way to our Free Software roadmap.

Let's share the software for a better Free Software !

happy hacking !